lifesciences/mental-health-counseling/scripts/requirements-aws.txt
aws-cdk-lib==2.207.0
boto3==1.39.14